chore(global): обновил версию vega-ui и подтянул ченджы с ветки 2733
Jira issue: VEGA-2778
